Replace wercker with Travis CI
This puts the rhumb set up inline with nap. We may want to look
into sharing Makefile and other setups across libraries like these,
to reduce duplication of effort and what not, but it's too soon
to call now.

I removed wercker for two reasons: to use the same config as nap,
but also because as long as wercker doesn't have a good story on
working *with* Docker tooling without having to do a bunch of
ceremony around it (see wercker/support#131) it's just not worth
the hassle for this kind of project.