introduce interpolation functionality to project
This enables Rhumb to interpolate params to produce paths.

This does not do anything with encoding or decoding the values of
params, so params with "/" and "?" characters in the values will produce
paths that the user may not expect.  This is out of scope for this PR,
though it will definitely be fixed later.